refactor: implement filterable builder macros for endpoint request builders
