+++ src/client.rs
@@ -299,58 +299,6 @@ impl OpenFIGIClient {
         OpenFIGIRequestBuilder::new(self.clone(), method, path)
     }
 
-    /// Parses a list of HTTP responses with comprehensive OpenFIGI-specific error handling.
-    pub(crate) async fn parse_list_response<T: DeserializeOwned>(
-        &self,
-        response: reqwest::Response,
-    ) -> Result<Vec<Result<T>>> {
-        let status = response.status();
-
-        // Early return for success case to optimize the common path
-        if status.is_success() {
-            // Deserialize the response body into the expected type `T`
-            let parsed_list: Vec<ResponseResult<T>> =
-                response.json().await.map_err(OpenFIGIError::from)?;
-
-            // Transform the parsed list into a `Result<T, OpenFIGIError>`.
-            let results: Vec<Result<T>> = parsed_list
-                .into_iter()
-                .map(|item| match item {
-                    ResponseResult::Success(data) => Ok(data),
-                    ResponseResult::Error(err) => Err(OpenFIGIError::response_error(
-                        status,
-                        format!("OpenFIGI API error: {}", err.error),
-                        String::new(),
-                    )),
-                })
-                .collect();
-
-            return Ok(results);
-        }
-
-        // For error cases, we need the URL
-        let url = response.url().clone();
-
-        // Rate-Limit-Info nur bei 429 extrahieren
-        let rate_limit_info = if status == reqwest::StatusCode::TOO_MANY_REQUESTS {
-            Self::extract_rate_limit_info(response.headers())
-        } else {
-            None
-        };
-
-        // Handle different HTTP status codes with OpenFIGI-specific context
-        let error_message = Self::format_error_message(status, &url, rate_limit_info);
-
-        // Get response text for error cases
-        let resp_text = response.text().await.map_err(OpenFIGIError::from)?;
-
-        Err(OpenFIGIError::response_error(
-            status,
-            error_message,
-            resp_text,
-        ))
-    }
-
     /// Parses single HTTP responses with comprehensive OpenFIGI-specific error handling.
     ///
     /// This internal method processes HTTP responses from the OpenFIGI API, providing
@@ -409,7 +357,63 @@ impl OpenFIGIClient {
             }
         }
 
-        // For error cases, we need the URL
+        return Err(self.handle_error_response(response).await);
+    }
+
+    /// Parses a list of HTTP responses from the OpenFIGI API, handling both successes and errors for batch mapping requests.
+    ///
+    /// This internal method processes HTTP responses that return a list of results (such as batch mapping requests).
+    /// Each item in the returned vector is a `Result<T, OpenFIGIError>`, where:
+    /// - `Ok(T)` contains a successfully parsed mapping result
+    /// - `Err(OpenFIGIError)` contains a detailed error for a failed mapping request
+    ///
+    /// # Arguments
+    ///
+    /// * `response` - The HTTP response from the OpenFIGI API
+    ///
+    /// # Returns
+    ///
+    /// Returns `Ok(Vec<Result<T, OpenFIGIError>>)` for successful HTTP responses, where each item is either a success or error for the corresponding batch request.
+    /// Returns `Err(OpenFIGIError)` if the HTTP response itself is an error (e.g., network failure, invalid status code).
+    ///
+    /// # Error Handling
+    ///
+    /// - Each batch item is parsed and mapped to either a success or error variant.
+    /// - If the HTTP response is not successful, a single `OpenFIGIError` is returned for the entire batch.
+    pub(crate) async fn parse_list_response<T: DeserializeOwned>(
+        &self,
+        response: reqwest::Response,
+    ) -> Result<Vec<Result<T>>> {
+        let status = response.status();
+
+        // Early return for success case to optimize the common path
+        if response.status().is_success() {
+            // Deserialize the response body into the expected type `T`
+            let parsed_list: Vec<ResponseResult<T>> =
+                response.json().await.map_err(OpenFIGIError::from)?;
+
+            // Transform the parsed list into a `Result<T, OpenFIGIError>`.
+            let results: Vec<Result<T>> = parsed_list
+                .into_iter()
+                .map(|item| match item {
+                    ResponseResult::Success(data) => Ok(data),
+                    ResponseResult::Error(err) => Err(OpenFIGIError::response_error(
+                        status,
+                        format!("OpenFIGI API error: {}", err.error),
+                        String::new(),
+                    )),
+                })
+                .collect();
+
+            return Ok(results);
+        }
+
+        return Err(self.handle_error_response(response).await);
+    }
+
+    /// Handles non-successful HTTP responses by creating a detailed `OpenFIGIError`.
+    async fn handle_error_response(&self, response: reqwest::Response) -> OpenFIGIError {
+        let status = response.status();
         let url = response.url().clone();
 
         // Rate-Limit-Info nur bei 429 extrahieren
@@ -422,14 +426,10 @@ impl OpenFIGIClient {
         // Handle different HTTP status codes with OpenFIGI-specific context
         let error_message = Self::format_error_message(status, &url, rate_limit_info);
 
-        // Get response text for error cases
-        let resp_text = response.text().await.map_err(OpenFIGIError::from)?;
+        // Use `unwrap_or_default` to avoid panics if text cannot be read
+        let resp_text = response.text().await.unwrap_or_default();
 
-        Err(OpenFIGIError::response_error(
-            status,
-            error_message,
-            resp_text,
-        ))
+        OpenFIGIError::response_error(status, error_message, resp_text)
     }
 
     /// Extracts rate limit information from HTTP response headers.
