+++ src/endpoint/mod.rs
@@ -9,7 +9,7 @@
 //! - [`crate::endpoint::filter`] - Filter endpoint for finding FIGIs using key words and other filters. The results are listed alphabetically by FIGI and include the number of results.
 //! - [`crate::endpoint::mapping`] - Mapping endpoint for converting third party identifiers to FIGIs
 //! - [`crate::endpoint::search`] - Search endpoint for finding FIGIs using key words and other filters.
-
 pub mod filter;
+mod macros;
 pub mod mapping;
 pub mod search;
